trie/epochmeta: add bloom filter, opt lock;
