+++ trie/epochmeta/database.go
@@ -1,13 +1,10 @@
 package epochmeta
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
-	"math/big"
-	"sync"
-
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/metrics"
+	"math/big"
 
 	"github.com/ethereum/go-ethereum/rlp"
 
@@ -17,6 +14,13 @@ import (
 
 const (
 	AccountMetadataPath = "m"
+)
+
+var (
+	metaAccessMeter       = metrics.NewRegisteredMeter("epochmeta/access", nil)
+	metaHitDiffMeter      = metrics.NewRegisteredMeter("epochmeta/access/hit/diff", nil)
+	metaHitDiskCacheMeter = metrics.NewRegisteredMeter("epochmeta/access/hit/diskcache", nil)
+	metaHitDiskMeter      = metrics.NewRegisteredMeter("epochmeta/access/hit/disk", nil)
 )
 
 type BranchNodeEpochMeta struct {
@@ -45,106 +49,28 @@ func DecodeFullNodeEpochMeta(enc []byte) (*BranchNodeEpochMeta, error) {
 	return &n, nil
 }
 
-// TODO(0xbundler): modify it as reader
-type Storage interface {
-	Get(addr common.Hash, path string) ([]byte, error)
-	Delete(addr common.Hash, path string) error
-	Put(addr common.Hash, path string, val []byte) error
-	Commit(number *big.Int, blockRoot common.Hash) error
+type Reader struct {
+	snap snapshot
+	tree *SnapshotTree
 }
 
-type StorageRW struct {
-	snap    snapshot
-	tree    *SnapshotTree
-	dirties map[common.Hash]map[string][]byte
-
-	stale bool
-	lock  sync.RWMutex
-}
-
-// NewEpochMetaDatabase first find snap by blockRoot, if got nil, try using number to instance a read only storage
-func NewEpochMetaDatabase(tree *SnapshotTree, number *big.Int, blockRoot common.Hash) (Storage, error) {
+// NewReader first find snap by blockRoot, if got nil, try using number to instance a read only storage
+func NewReader(tree *SnapshotTree, number *big.Int, blockRoot common.Hash) (*Reader, error) {
 	snap := tree.Snapshot(blockRoot)
 	if snap == nil {
 		// try using default snap
 		if snap = tree.Snapshot(types.EmptyRootHash); snap == nil {
 			return nil, fmt.Errorf("cannot find target epoch layer %#x", blockRoot)
 		}
-		log.Debug("NewEpochMetaDatabase use default database", "number", number, "root", blockRoot)
+		log.Debug("NewReader use default database", "number", number, "root", blockRoot)
 	}
-	return &StorageRW{
-		snap:    snap,
-		tree:    tree,
-		dirties: make(map[common.Hash]map[string][]byte),
+	return &Reader{
+		snap: snap,
+		tree: tree,
 	}, nil
 }
 
-func (s *StorageRW) Get(addr common.Hash, path string) ([]byte, error) {
-	s.lock.RLock()
-	defer s.lock.RUnlock()
-	// TODO(0xbundler): remove cache
-	sub, exist := s.dirties[addr]
-	if exist {
-		if val, ok := sub[path]; ok {
-			return val, nil
-		}
-	}
-
-	// TODO(0xbundler): metrics hit count
+func (s *Reader) Get(addr common.Hash, path string) ([]byte, error) {
+	metaAccessMeter.Mark(1)
 	return s.snap.EpochMeta(addr, path)
-}
-
-func (s *StorageRW) Delete(addr common.Hash, path string) error {
-	s.lock.RLock()
-	defer s.lock.RUnlock()
-	if s.stale {
-		return errors.New("storage has staled")
-	}
-	_, ok := s.dirties[addr]
-	if !ok {
-		s.dirties[addr] = make(map[string][]byte)
-	}
-
-	s.dirties[addr][path] = nil
-	return nil
-}
-
-func (s *StorageRW) Put(addr common.Hash, path string, val []byte) error {
-	prev, err := s.Get(addr, path)
-	if err != nil {
-		return err
-	}
-	if bytes.Equal(prev, val) {
-		return nil
-	}
-
-	s.lock.RLock()
-	defer s.lock.RUnlock()
-	if s.stale {
-		return errors.New("storage has staled")
-	}
-
-	_, ok := s.dirties[addr]
-	if !ok {
-		s.dirties[addr] = make(map[string][]byte)
-	}
-	s.dirties[addr][path] = val
-	return nil
-}
-
-// Commit if you commit to an unknown parent, like deeper than 128 layers, will get error
-func (s *StorageRW) Commit(number *big.Int, blockRoot common.Hash) error {
-	s.lock.Lock()
-	defer s.lock.Unlock()
-	if s.stale {
-		return errors.New("storage has staled")
-	}
-
-	s.stale = true
-	err := s.tree.Update(s.snap.Root(), number, blockRoot, s.dirties)
-	if err != nil {
-		return err
-	}
-
-	return s.tree.Cap(blockRoot)
 }
